+++ src/cryptoUtils/unitConverter.ts
@@ -2,7 +2,7 @@ import web3 from 'web3';
 
 export function fullOrbsFromWeiOrbs(weiOrbs?: bigint): number {
   if (!weiOrbs) return 0;
-  return parseInt(web3.utils.fromWei(weiOrbs.toString(), 'ether'));
+  return parseFloat(web3.utils.fromWei(weiOrbs.toString(), 'ether'));
 }
 
 export function weiOrbsFromFullOrbs(fullOrbs?: number): bigint {
